Refactor block definition structure & data loading
