+++ bundles/org.openhab.ui/web/src/assets/definitions/blockly/_tmp/blocks-http.js
