+++ bundles/org.openhab.ui/web/src/assets/definitions/blockly/blocks-valuestorage.js
