+++ bundles/org.openhab.ui/web/src/assets/definitions/blockly/fields/audiosink-field.js
