+++ bundles/org.openhab.ui/web/src/assets/definitions/blockly/fields/item-field.js
