+++ bundles/org.openhab.ui/web/src/assets/definitions/blockly/fields/thing-field.js
